Drop redundant "NOT NULL" constraint on primary key column
Primary keys have implicit not null constraints.  (Except in SQLite,
apparently, but we don't really care about that.  Constraints are not
supposed to be violated ever, so no huge issue that they are missing.)

The reason for this change is to make the SQL more uniform, which makes
automatic translation of non-standard features (like "SERIAL" here)
easier.