Read at most environ["CONTENT_LENGTH"] bytes of request body
This is actually required by PEP 333:

  The server is not required to read past the client's specified
  Content-Length, and is allowed to simulate an end-of-file condition
  if the application attempts to read past that point.  The application
  should not attempt to read more data than is specified by the
  CONTENT_LENGTH variable.

Since mod_wsgi simulates end-of-file, our old behavior worked just fine.
But wsgiref doesn't, and instead blocks until the browser closes the
connection (which naturally makes it hard to write a response back,) so
this more correct behavior is required for compatibility with wsgiref.