Rewrite various queries to avoid non-standard DELETE/UPDATE syntax
The "DELETE FROM ... USING ..." and "UPDATE ... FROM ..." syntax
extensions for doing joins in DELETE and UPDATE statements is another
PostgreSQL:ism.  A rather convenient one, but quite easy to avoid in
most cases.

I've confirmed that PostgreSQL's query planner comes up with exactly the
same strategy in the two cases

  DELETE FROM x USING y WHERE x.a=y.a and y.b

and

  DELETE FROM x WHERE a IN (SELECT a FROM y WHERE b);

which should mean that there's no performance drawback to use the
seemingly more complicated sub-query variant.

In some cases the rewrite splits a query into two, one that selects
from 'y' and one that deletes from 'x' using cursor.executemany() and
the results from the first query.  This ought to be slightly slower,
since it's two queries instead of one and since the data is actually
transferred from the database server and back.  But in the cases where
such a split was necessary, there's not expected to be a huge amount
of data to transfer, and the performance impact is expected to be of
little significance.