Improve caching of static resources handled internally
Static resources (stylesheets and scripts) referenced from generated HTML
pages have a base-36 encoded timestamp in the query string, which means
they can essentially be cached forever.  This is implemented when Apache
handles the resources via mod_expires, so handle it here as well.

This handling is actually a bit more accurate, setting the generous cache
control headers only when the URL contains a (correct) timestamp.  This
means sub-resources (i.e. images referenced from stylesheets) are not
affected, since their URLs do not have the timestamp.  This is of course
a bit unfortunate; we want them to be cached as well.